e2e: Extract viem initialization code into viem_setup.mjs for reuse (#261)

e2e(js): Add mainnet & baklava config in e2e viem setup (#331, #340)
